Module aws_sdk_dynamodbstreams::model

Modules

identity

See Identity

key_schema_element

See KeySchemaElement

record

See Record

sequence_number_range

See SequenceNumberRange

shard

See Shard

stream

See Stream

stream_description

See StreamDescription

stream_record

See StreamRecord

Structs

Identity

Contains details about the type of identity that made the request.

KeySchemaElement

Represents a single element of a key schema. A key schema specifies

Record

A description of a unique event within a stream.

SequenceNumberRange

The beginning and ending sequence numbers for the stream records contained within a shard.

Shard

A uniquely identified group of stream records within a stream.

Stream

Represents all of the data describing a particular stream.

StreamDescription

Represents all of the data describing a particular stream.

StreamRecord

A description of a single data modification that was performed on an item in a DynamoDB table.

Enums

AttributeValue

KeyType

OperationType

ShardIteratorType

StreamStatus

StreamViewType